sensor/requirements.txt
paho-mqtt==1.5.0
npyscreen==4.10.5
